BAU: Move Builder classes for saml-domain-objects into saml-domain-objects
These classes are all just Builders of classes defined in saml-domain-objects.
Previously these builder classes lived in saml-test-utils.
This created an unnessesary dependency between the two libraries.
Moving the classes into this library moves them closer to the objects that
they relate to, and allows us to break the dependency of saml-test-utils
on saml-domain-objects.

I have also marked all of the moved classes as deprecated, as they serve
only to build deprecated classes.

Solo: @michaelwalker